Add C++ solution for Digit Queries

+++ Introductory_Problems/Digit_Queries/Digit_Queries.cpp
@@ -36,7 +36,7 @@ int main() {
     int q;
     cin>>q;
     while(q--){
-        int k;
+        ll k;
         cin>>k;
         cout<<solve(k)<<"\n";
     }
